+++ modules/hanna-react/src/FocusTrap.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import { focusElement } from '@reykjavik/hanna-utils';
 
 type FocusableElement = Element & { focus: () => void };
 
@@ -10,7 +11,10 @@ export type FocusTrapProps = {
   /**
    * How deep the trap is placed in the DOM tree beneath its container element.
    *
-   * Default: `1`
+   * This is useful when the trap needs to be placed inside an element that's
+   * not a direct child of the desired trapping container.
+   *
+   * Default: `1`  (i.e. use the parent element.)
    */
   depth?: number;
 };
@@ -39,7 +43,7 @@ export const FocusTrap = (props: FocusTrapProps) => {
           return;
         }
         const focusables = container.querySelectorAll<FocusableElement>(
-          'a[href], input, select, textarea, button, [tabindex]:not(.FocusTrap):not([tabindex="-1"])'
+          focusElement.keyboardFocusableSelector
         );
         const delta = props.atTop ? -1 : 1;
         let i = delta < 0 ? focusables.length - 1 : 0;