+++ modules/hanna-react/src/FormField.tsx
@@ -121,7 +121,7 @@ type FormFieldProps = FormFieldGroupWrappingProps & {
 
   renderInput(
     className: InputClassNames,
-    inputProps: FormFieldInputProps,
+    inputProps: FormFieldInputProps & { id: string },
     addFocusProps: FocusPropMaker,
     isBrowser?: boolean
   ): JSX.Element;
@@ -211,7 +211,7 @@ export const FormField = (props: FormFieldProps) => {
     </abbr>
   );
 
-  const inputProps: FormFieldInputProps = {
+  const inputProps: FormFieldInputProps & { id: string } = {
     id: domid,
     disabled: disabled,
     required: required,