+++ modules/hanna-react/src/utils/a11yHelpers.ts
@@ -1,13 +1,15 @@
-import { MouseEventHandler } from 'react';
+import { MouseEvent as ReactMouseEvent } from 'react';
 import { focusElement } from '@reykjavik/hanna-utils';
 
-export const handleAnchorLinkClick: MouseEventHandler<HTMLAnchorElement> = (e) => {
+export const handleAnchorLinkClick = (
+  e: ReactMouseEvent<HTMLAnchorElement>,
+  forwardToFirstFocusable?: boolean
+) => {
   e.preventDefault();
   const targetId = e.currentTarget.hash.slice(1);
   const targetElm = targetId && document.getElementById(targetId);
   if (!targetElm) {
     return;
   }
-  targetElm.tabIndex = -1;
-  focusElement(targetElm);
+  focusElement(targetElm, forwardToFirstFocusable);
 };