+++ modules/hanna-sprinkles/src/Tabs.tsx
@@ -2,10 +2,9 @@ import './_/initHannaNamespace.js';
 
 import React, { useMemo, useState } from 'react';
 import ReactDOM from 'react-dom';
-import domid from '@hugsmidjan/qj/domid';
 import { getFrag } from '@hugsmidjan/qj/frag';
 import { TabItemProps, Tabs } from '@reykjavik/hanna-react/Tabs';
-import { WrapperElmProps } from '@reykjavik/hanna-react/utils';
+import { domid, WrapperElmProps } from '@reykjavik/hanna-react/utils';
 import { notNully } from '@reykjavik/hanna-utils';
 
 import { autoSeenEffectsRefresh, autoSeenEffectWrapperProps } from './_/addSeenEffect.js';