+++ modules/hanna-utils/CHANGELOG.md
@@ -3,6 +3,17 @@
 ## Upcoming...
 
 - ... <!-- Add new lines here. -->
+
+## 0.2.19
+
+_2024-11-21_
+
+- feat: Add property `focusElement.keyboardFocusableSelector`
+
+## 0.2.18
+
+_2024-11-20_
+
 - feat: Add testing type `NotExpect`
 
 ## 0.2.17